requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
requests==2.31.0
beautifulsoup4==4.12.2
pillow==10.1.0
flask==2.3.3
flask-cors==4.0.0